GEODE-9642: wait for colocation to complete as a part of create missing buckets task
