+++ geode-core/src/main/java/org/apache/geode/internal/cache/PRHARedundancyProvider.java
@@ -145,13 +145,6 @@ public class PRHARedundancyProvider {
 
   private static final ThreadLocal<Boolean> forceLocalPrimaries = new ThreadLocal<>();
 
-  private static final int MAX_NUMBER_INTERVALS = 31;
-
-  private static final int SMALL_INTERVALS = 5;
-  private static final int MEDIUM_INTERVALS = 10;
-  private static final int LARGE_INTERVALS = 15;
-
-
   @MakeNotStatic
   private static final Long DATASTORE_DISCOVERY_TIMEOUT_MILLISECONDS =
       Long.getLong(DATASTORE_DISCOVERY_TIMEOUT_PROPERTY_NAME);
@@ -1729,9 +1722,7 @@ public class PRHARedundancyProvider {
   }
 
   void scheduleCreateMissingBuckets() {
-    if (partitionedRegion.getColocatedWith() != null &&
-        waitForColocationCompleted(partitionedRegion)) {
-
+    if (partitionedRegion.getColocatedWith() != null) {
       Runnable task = new CreateMissingBucketsTask(this);
       final InternalResourceManager resourceManager =
           partitionedRegion.getGemFireCache().getInternalResourceManager();
@@ -1900,42 +1891,6 @@ public class PRHARedundancyProvider {
       getPersistentBucketRecoverer().await();
     }
   }
-
-
-  /**
-   * Wait for Colocation to complete. Wait all nodes to Register this PartitionedRegion.
-   */
-  private boolean waitForColocationCompleted(PartitionedRegion partitionedRegion) {
-    int retryCount = 0;
-    int sleepInterval = PartitionedRegionHelper.DEFAULT_WAIT_PER_RETRY_ITERATION;
-
-    while (!ColocationHelper.isColocationComplete(partitionedRegion)
-        && (retryCount < MAX_NUMBER_INTERVALS)) {
-
-      // Didn't time out. Sleep a bit and then continue
-      boolean interrupted = Thread.interrupted();
-      try {
-        Thread.sleep(sleepInterval);
-      } catch (InterruptedException ignore) {
-        interrupted = true;
-      } finally {
-        if (interrupted) {
-          Thread.currentThread().interrupt();
-        }
-      }
-      retryCount++;
-      if (retryCount == SMALL_INTERVALS) {
-        sleepInterval = 2 * PartitionedRegionHelper.DEFAULT_WAIT_PER_RETRY_ITERATION;
-      } else if (retryCount == MEDIUM_INTERVALS) {
-        sleepInterval = 5 * PartitionedRegionHelper.DEFAULT_WAIT_PER_RETRY_ITERATION;
-      } else if (retryCount == LARGE_INTERVALS) {
-        sleepInterval = 10 * PartitionedRegionHelper.DEFAULT_WAIT_PER_RETRY_ITERATION;
-      }
-    }
-    return ColocationHelper.isColocationComplete(partitionedRegion);
-
-  }
-
 
   public boolean isPersistentRecoveryComplete() {
     if (!checkMembersColocation(partitionedRegion, partitionedRegion.getMyId())) {
